Remove Application Credentials
This is now performed in Unikorn Server, so there's no need to expose
this to clients (wooo!)  However... this led me to looking at the client
code, which is Unikorn's codebase, so why not resuse and wield the power
of `go get`?  But that led me to other... horrors?  The good news is
this repo is now 400,000 lines lighter, and also uses modern go modules!